+++ packages/directives/src/index.ts
@@ -27,5 +27,6 @@ export { default as clickoutside } from './clickoutside';
 export { default as bkEllipsis, createInstance as bkEllipsisInstance } from './ellipsis';
 export { default as mousewheel } from './mousewheel';
 export { default as bkTooltips } from './tooltips';
+export { vBkloading as bkLoading } from '@bkui-vue/loading';
 export { overflowTitle } from '@bkui-vue/overflow-title';
 // export { createInstance as bkEllipsisInstance } from './ellipsis';