+++ packages/table/src/plugins/head-filter.tsx
@@ -33,7 +33,6 @@ import { Funnel } from '@bkui-vue/icon';
 import Input from '@bkui-vue/input';
 import Popover from '@bkui-vue/popover';
 import { classes, PropTypes, RenderType } from '@bkui-vue/shared';
-import VirtualRender from '@bkui-vue/virtual-render';
 
 import { Column, IColumnType, IFilterShape } from '../props';
 
@@ -134,6 +133,12 @@ export default defineComponent({
 
       return defaultMin;
     });
+
+    const contentStyle = computed(() => ({
+      maxHeight: `${maxHeight.value}px`,
+      minHeight: `${minHeight.value}px`,
+      height: /%$/.test(`${height.value}`) ? height.value : `${height.value}px`,
+    }));
 
     const getRegExp = (val: boolean | number | string, flags = 'ig') =>
       new RegExp(`${val}`.replace(/[-[\]{}()*+?.,\\^$|#\s]/g, '\\$&'), flags);
@@ -282,8 +287,8 @@ export default defineComponent({
     };
 
     const renderFilterList = scope => {
-      if (scope.data.length) {
-        return scope.data.map((item: { value: string; text: string; tipKey?: string }) => (
+      if (scope.length) {
+        return scope.map((item: { value: string; text: string; tipKey?: string }) => (
           <div
             key={item.value}
             ref={filterPopoverRef}
@@ -330,21 +335,12 @@ export default defineComponent({
                 <Input v-model={searchValue.value}></Input>
               </div>
               <BkCheckboxGroup class='content-list'>
-                <VirtualRender
-                  ref={refVirtualRender}
-                  height={height.value}
-                  className='content-items'
-                  lineHeight={32}
-                  list={localData.value}
-                  maxHeight={maxHeight.value}
-                  minHeight={minHeight.value}
-                  scrollEvent={true}
-                  throttleDelay={0}
+                <div
+                  style={contentStyle.value}
+                  class='content-items'
                 >
-                  {{
-                    default: renderFilterList,
-                  }}
-                </VirtualRender>
+                  {renderFilterList(localData.value)}
+                </div>
               </BkCheckboxGroup>
               <div class='content-footer'>
                 {renderSaveBtn()}